refactor:StationResult 및 관련 타입을 types/station.ts로 분리
